Introduce spooles error handling
Add the possibility to hook in an error handling function. This may mean
that C++ users pass in a function that throws an exception, others may
prefer having their own function with setjmp/longjmp or a brakepoint, or
want to dump a stack trace.

For users of spooles, this introduces the function
registerExitFunction(), which takes a pointer to a function that can be
used to add some user defined behavior. Since spooles has run into an
unrecoverable error, the function should probably make sure to exit(),
throw, or longjmp to prevent any crashes from occurring.

All exit() calls are replaced with spoolesFatal() in the next commit.